Add cognitive complexity
