+++ sonarts-core/src/rules/cognitiveComplexityRule.ts
@@ -20,7 +20,8 @@
 import * as tslint from "tslint";
 import * as ts from "typescript";
 import { SonarRuleMetaData } from "../sonarRule";
-import { is, functionLikeMainToken, drillDownThroughParenthesis } from "../utils/navigation";
+import { functionLikeMainToken } from "../utils/navigation";
+import { FunctionCollector } from "../utils/cognitiveComplexity";
 
 export class Rule extends tslint.Rules.AbstractRule {
   public static metadata: SonarRuleMetaData = {
@@ -45,234 +46,41 @@ export class Rule extends tslint.Rules.AbstractRule {
   }
 
   public apply(sourceFile: ts.SourceFile): tslint.RuleFailure[] {
-    return this.applyWithWalker(new FunctionWalker(sourceFile, this.getOptions(), this.threshold));
+    const walker = new tslint.RuleWalker(sourceFile, this.getOptions());
+
+    const functionCollector = new FunctionCollector();
+    functionCollector.walk(sourceFile);
+
+    functionCollector.functionComplexities.forEach(functionComplexity => {
+      if (functionComplexity.complexity > this.threshold) {
+        walker.addFailureAtNode(
+          functionLikeMainToken(functionComplexity.functionNode),
+          getMessage(functionComplexity.functionNode, functionComplexity.complexity, this.threshold),
+        );
+      }
+    });
+
+    return walker.getFailures();
   }
 
   public static DEFAULT_THRESHOLD = 15;
-
-  public static TARGETED_KINDS = [
-    // ArrowFunctions and FunctionExpressions are included in computation
-    ts.SyntaxKind.FunctionDeclaration,
-    ts.SyntaxKind.MethodDeclaration,
-    ts.SyntaxKind.Constructor,
-    ts.SyntaxKind.GetAccessor,
-    ts.SyntaxKind.SetAccessor,
-  ];
 }
 
-class FunctionWalker extends tslint.RuleWalker {
-  threshold: number;
-
-  constructor(sourceFile: ts.SourceFile, options: tslint.IOptions, threshold: number) {
-    super(sourceFile, options);
-    this.threshold = threshold;
-  }
-
-  public visitNode(node: ts.Node): void {
-    if (is(node, ...Rule.TARGETED_KINDS)) {
-      const body = (node as ts.FunctionLikeDeclarationBase).body as ts.Node;
-      if (body) {
-        const complexity = this.getComplexity(body);
-        if (complexity > this.threshold) {
-          this.addFailureAtNode(
-            functionLikeMainToken(node as ts.FunctionLikeDeclaration),
-            FunctionWalker.getMessage(node, complexity, this.threshold),
-          );
-        }
-      }
-    }
-    super.visitNode(node);
-  }
-
-  private getComplexity(node: ts.Node): number {
-    let complexityWalker = new ComplexityWalker();
-    complexityWalker.walk(node);
-    return complexityWalker.getComplexity();
-  }
-
-  private static getMessage(node: ts.Node, complexity: number, threshold: number): string {
-    const functionName = FunctionWalker.getFunctionName(node);
-    return `Refactor this ${functionName} to reduce its Cognitive Complexity from ${complexity} to the ${threshold} allowed.`;
-  }
-
-  private static getFunctionName(node: ts.Node): string {
-    switch (node.kind) {
-      case ts.SyntaxKind.MethodDeclaration:
-        return "method";
-      case ts.SyntaxKind.Constructor:
-        return "constructor";
-      case ts.SyntaxKind.GetAccessor:
-      case ts.SyntaxKind.SetAccessor:
-        return "accessor";
-      default:
-        return "function";
-    }
-  }
+function getMessage(node: ts.Node, complexity: number, threshold: number): string {
+  const functionName = getFunctionName(node);
+  return `Refactor this ${functionName} to reduce its Cognitive Complexity from ${complexity} to the ${threshold} allowed.`;
 }
 
-class ComplexityWalker extends tslint.SyntaxWalker {
-  private complexity = 0;
-  private nesting = 0;
-  private logicalOperationsToIgnore: ts.BinaryExpression[] = [];
-
-  public getComplexity(): number {
-    return this.complexity;
-  }
-
-  public visitIfStatement(node: ts.IfStatement): void {
-    if (this.isElseIf(node)) {
-      this.addComplexityWithoutNesting();
-    } else {
-      this.addComplexityWithNesting();
-    }
-
-    this.walk(node.expression);
-    this.visitWithNesting(node.thenStatement);
-
-    if (node.elseStatement) {
-      if (is(node.elseStatement, ts.SyntaxKind.IfStatement)) {
-        this.walk(node.elseStatement);
-      } else {
-        this.addComplexityWithoutNesting();
-        this.visitWithNesting(node.elseStatement);
-      }
-    }
-  }
-
-  private isElseIf(node: ts.IfStatement): boolean {
-    if (node.parent) {
-      return is(node.parent, ts.SyntaxKind.IfStatement) && node === (node.parent as ts.IfStatement).elseStatement;
-    }
-    return false;
-  }
-
-  public visitSwitchStatement(node: ts.SwitchStatement): void {
-    this.addComplexityWithNesting();
-    this.walk(node.expression);
-    this.visitWithNesting(node.caseBlock);
-  }
-
-  public visitClassDeclaration(_node: ts.ClassDeclaration) {
-    // do nothing, skip inner classes - its methods complexity will be computed indivudally
-  }
-
-  public visitFunctionDeclaration(_node: ts.FunctionDeclaration): void {
-    // do nothing, skip inner functions - their own complexity will be computed
-  }
-
-  public visitFunctionExpression(node: ts.FunctionExpression): void {
-    this.addComplexityWithNesting();
-    this.visitWithNesting(node.body);
-  }
-
-  public visitArrowFunction(node: ts.ArrowFunction): void {
-    this.addComplexityWithNesting();
-    this.visitWithNesting(node.body);
-  }
-
-  public visitContinueStatement(node: ts.ContinueStatement): void {
-    if (node.label) {
-      this.addComplexityWithoutNesting();
-    }
-    super.visitContinueStatement(node);
-  }
-
-  public visitBreakStatement(node: ts.BreakStatement): void {
-    if (node.label) {
-      this.addComplexityWithoutNesting();
-    }
-    super.visitBreakStatement(node);
-  }
-
-  public visitCatchClause(node: ts.CatchClause): void {
-    this.addComplexityWithNesting();
-    this.visitWithNesting(node.block);
-  }
-
-  public visitForStatement(node: ts.ForStatement): void {
-    this.visitLoop(node.statement, node.initializer, node.condition, node.incrementor);
-  }
-
-  public visitForInStatement(node: ts.ForInStatement): void {
-    this.visitLoop(node.statement, node.initializer, node.expression);
-  }
-
-  public visitForOfStatement(node: ts.ForOfStatement): void {
-    this.visitLoop(node.statement, node.initializer, node.expression);
-  }
-
-  public visitDoStatement(node: ts.DoStatement): void {
-    this.visitLoop(node.statement, node.expression);
-  }
-
-  public visitWhileStatement(node: ts.WhileStatement): void {
-    this.visitLoop(node.statement, node.expression);
-  }
-
-  private visitLoop(nestedNode: ts.Node, ...notNestedExpressions: (ts.Node | undefined)[]) {
-    this.addComplexityWithNesting();
-    this.walkNodes(notNestedExpressions);
-    this.visitWithNesting(nestedNode);
-  }
-
-  public visitConditionalExpression(node: ts.ConditionalExpression): void {
-    this.addComplexityWithNesting();
-    this.walk(node.condition);
-    this.visitWithNesting(node.whenTrue);
-    this.visitWithNesting(node.whenFalse);
-  }
-
-  public visitBinaryExpression(node: ts.BinaryExpression): void {
-    const operator = node.operatorToken;
-    if (is(operator, ts.SyntaxKind.AmpersandAmpersandToken, ts.SyntaxKind.BarBarToken)) {
-      const leftChild = drillDownThroughParenthesis(node.left);
-      const rightChild = drillDownThroughParenthesis(node.right);
-
-      const leftChildOfSameKind = ComplexityWalker.isOperator(leftChild, operator.kind);
-      const rightChildOfSameKind = ComplexityWalker.isOperator(rightChild, operator.kind);
-
-      // move always to the leftmost operator
-      if (rightChildOfSameKind) {
-        this.logicalOperationsToIgnore.push(rightChild as ts.BinaryExpression);
-      }
-
-      if (!this.isIgnoredOperation(node) && !leftChildOfSameKind) {
-        this.addComplexityWithoutNesting();
-      }
-    }
-    super.visitBinaryExpression(node);
-  }
-
-  private static isOperator(node: ts.Node, operatorKind: ts.BinaryOperator): boolean {
-    if (is(node, ts.SyntaxKind.BinaryExpression)) {
-      return is((node as ts.BinaryExpression).operatorToken, operatorKind);
-    }
-    return false;
-  }
-
-  private isIgnoredOperation(node: ts.BinaryExpression): boolean {
-    return this.logicalOperationsToIgnore.indexOf(node) > -1;
-  }
-
-  private walkNodes(nodes: (ts.Node | undefined)[]) {
-    for (const node of nodes) {
-      if (node) {
-        this.walk(node);
-      }
-    }
-  }
-
-  private visitWithNesting(node: ts.Node): void {
-    this.nesting++;
-    this.walk(node);
-    this.nesting--;
-  }
-
-  private addComplexityWithoutNesting(): void {
-    this.complexity += 1;
-  }
-
-  private addComplexityWithNesting(): void {
-    this.complexity += this.nesting + 1;
+function getFunctionName(node: ts.Node): string {
+  switch (node.kind) {
+    case ts.SyntaxKind.MethodDeclaration:
+      return "method";
+    case ts.SyntaxKind.Constructor:
+      return "constructor";
+    case ts.SyntaxKind.GetAccessor:
+    case ts.SyntaxKind.SetAccessor:
+      return "accessor";
+    default:
+      return "function";
   }
 }
