+++ sonarts-core/src/utils/cyclomaticComplexity.ts
